Allez les bleus

+++ src/test/kotlin/fr/insalyon/pld/agile/model/DistanceTest.kt
@@ -11,7 +11,7 @@ class DistanceTest {
         val d0  =  1L.dam
         Assert.assertEquals(10, d0.to(Distance.DistanceUnit.M).value)
         Assert.assertEquals(1,d0.to(Distance.DistanceUnit.DAM).value)
-        Assert.assertEquals(0.0.toInt(), d0.to(Distance.DistanceUnit.KM).value)
+        Assert.assertEquals(0, d0.to(Distance.DistanceUnit.KM).value)
 
         val d1 = 10L.km
         Assert.assertEquals(10000, d1.to(Distance.DistanceUnit.M).value)
